Add support for reading Tix and Mix Files
With the addition of reading Tix and Mix files, the command line
argument structure was altered slightly to account for the changes.

An overarching monad stack was also added for tracking the configuration
and reporting errors to the user.

Tests were updated (partially to try and generate a tix/mix combo to
use) instead the project Emulator6502 was used as our testdata.